Added documentation for the auth system
