+++ lib/src/services/authentication.dart
@@ -4,71 +4,92 @@ import "package:arcane_framework/arcane_framework.dart";
 import "package:flutter/widgets.dart";
 import "package:flutter_bloc/flutter_bloc.dart";
 
+part "../authentication/authentication_enums.dart";
+part "../authentication/authentication_interface.dart";
+
+/// Provides a standard interface to handle authentication-related tasks.
+///
+/// To get started, first ensure that an `ArcaneAuthInterface` has been
+/// registered.
 class ArcaneAuthenticationService extends ArcaneService {
   ArcaneAuthenticationService._internal();
 
   static bool _mocked = false;
   static final ArcaneAuthenticationService _instance =
       ArcaneAuthenticationService._internal();
+
+  /// Provides access to the singleton instance of this service.
   static ArcaneAuthenticationService get I => _instance;
 
   AuthenticationStatus _status = AuthenticationStatus.unauthenticated;
+
+  /// Returns the current `AuthenticationStatus`.
+  ///
+  /// Available values:
+  /// - `authenticated`: The user has successfully authenticated and is logged in.
+  /// - `unauthenticated`: The user has not yet logged in.
+  /// - `debug`: Debug mode has been enabled, enabling development features.
   AuthenticationStatus get status => _status;
 
   static late ArcaneAuthInterface _authInterface;
+
+  /// Provides direct access to the registered `ArcaneAuthInterface`, if one has
+  /// been registered.
   ArcaneAuthInterface get authInterface => _authInterface;
 
+  /// A shortcut to `status == AuthenticationStatus.authenticated`.
   bool get isAuthenticated => status == AuthenticationStatus.authenticated;
+
+  /// A shortcut to the `isSignedIn` getter of the registered `ArcaneAuthInterface`.
   Future<bool> get isSignedIn => authInterface.isSignedIn;
 
+  /// Returns a JWT access token if the registered `ArcaneAuthInterface`
+  /// provides one. This token is often used in the headers of HTTP requests
+  /// to the backend API.
   Future<String?> get accessToken => authInterface.accessToken;
+
+  /// Returns a JWT refresh token if the registered `ArcaneAuthInterface`
+  /// provides one.
   Future<String?> get refreshToken => authInterface.refreshToken;
 
   static ArcaneSecureStorage get _storage => Arcane.storage;
 
+  /// Registers an `ArcaneAuthInterface` within the `ArcaneAuthenticationService`.
   Future<void> registerInterface(ArcaneAuthInterface authInterface) async {
     _authInterface = authInterface;
     await authInterface.init();
   }
 
-  /// Sets [status] to [AuthenticationStatus.debug].
+  /// Sets `status` to `AuthenticationStatus.debug`. If `onDebugModeSet` has
+  /// been specified, the method will be triggered after the new status has been
+  /// set.
   Future<void> setDebug(
-    BuildContext context,
+    BuildContext context, {
     Future<void> Function()? onDebugModeSet,
-  ) async {
+  }) async {
     if (_mocked) return;
-
-    Arcane.logger.log(
-      "!!!!! DEBUG MODE ENABLED !!!!!",
-      level: Level.fatal,
-    );
 
     ArcaneEnvironment? environment;
 
     try {
       environment = context.read<ArcaneEnvironment>();
-
       await environment.enableDebugMode(onDebugModeSet);
     } catch (_) {
-      Arcane.logger.log(
-        "No ArcaneEnvironment found in BuildContext",
-        level: Level.error,
-      );
+      throw Exception("No ArcaneEnvironment found in BuildContext");
     }
 
     _setStatus(AuthenticationStatus.debug);
+    if (onDebugModeSet != null) await onDebugModeSet();
   }
 
-  /// Sets [status] to [AuthenticationStatus.authenticated].
+  /// Sets `status` to `AuthenticationStatus.authenticated`.
   void setAuthenticated() {
     if (_mocked) return;
 
     _setStatus(AuthenticationStatus.authenticated);
   }
 
-  /// Sets [status] to [AuthenticationStatus.unauthenticated] and triggers a
-  /// refresh in [AppRouter], effectively returning a logged out user to the
-  /// welcome screen.
+  /// Sets `status` to `AuthenticationStatus.unauthenticated`.
   void setUnauthenticated() {
     if (_mocked) return;
 
@@ -82,8 +103,8 @@ class ArcaneAuthenticationService extends ArcaneService {
     notifyListeners();
   }
 
-  /// Logs the current user out. Upon successful logout, [status] will be set to
-  /// [AuthenticationStatus.unauthenticated].
+  /// Logs the current user out. Upon successful logout, `status` will be set to
+  /// `AuthenticationStatus.unauthenticated`.
   Future<void> logOut({required VoidCallback onLoggedOut}) async {
     if (_mocked) return;
     if (status == AuthenticationStatus.unauthenticated) return;
@@ -93,34 +114,30 @@ class ArcaneAuthenticationService extends ArcaneService {
     await loggedOut.fold(
       onSuccess: (_) async {
         await _storage.deleteAll();
-
         setUnauthenticated();
-
-        Arcane.logger.log(
-          "Sign out completed successfully",
-          level: Level.info,
-        );
-
         onLoggedOut();
       },
       onError: (e) {
-        Arcane.logger.log(
-          "Error signing user out: $e",
-          level: Level.error,
-        );
+        throw Exception(e);
       },
     );
   }
 
-  /// Attempts to log in the user using their [email] and [password].
-  /// Upon successful login, [status] will be set to
-  /// [AuthenticationStatus.authenticated].
+  /// Attempts to log in the user using their `email` and `password`.
+  /// Upon successful login, `status` will be set to
+  /// `AuthenticationStatus.authenticated]` If `onLoggedIn` is specified, the
+  /// method will run after the authentication status has been updated.
+  /// When logging in with email and password, the user's email address will be
+  /// cached in `ArcaneSecureStorage`.
   Future<Result<void, String>> loginWithEmailAndPassword({
     required String email,
     required String password,
+    Future<void> Function()? onLoggedIn,
   }) async {
+    if (_mocked) return Result.ok(null);
+
     if (status != AuthenticationStatus.unauthenticated) {
-      return Result.error("Already signed in");
+      return Result.error("Cannot sign in. Status is already ${status.name}.");
     }
 
     final Result<void, String> result =
@@ -129,66 +146,38 @@ class ArcaneAuthenticationService extends ArcaneService {
       password: password,
     );
 
-    if (result.isFailure) {
-      Arcane.logger.log(
-        "Error signing in: ${result.error}",
-        level: Level.error,
-      );
-    }
-
     if (result.isSuccess) {
       await _storage.setValue(ArcaneSecureStorage.emailKey, email);
       setAuthenticated();
+      if (onLoggedIn != null) await onLoggedIn();
     }
 
     return result;
   }
 
-  /// Attempts to register a new account using the provided [email] and
-  /// [password]. Upon success, returns a [SignUpStep] indicating the next step
-  /// in the signup process.
+  /// Attempts to register a new account using the provided `email` and
+  /// `password`. Upon success, returns a `SignUpStep` indicating the next step
+  /// in the signup process as a `SignUpStep`.
   Future<Result<SignUpStep, String>> signup({
     required String email,
     required String password,
   }) async {
-    if (status == AuthenticationStatus.debug) {
-      return Result.error("Operation not supported in debug mode.");
-    }
-
+    if (_mocked) return Result.ok(SignUpStep.done);
     final Result<SignUpStep, String> result = await authInterface.signup(
       email: email,
       password: password,
     );
 
-    if (result.isFailure) {
-      Arcane.logger.log(
-        "Error signing up: ${result.error}",
-        level: Level.error,
-      );
-
-      return Result.error(result.error);
-    }
-
-    if (result.value == SignUpStep.confirmSignUp) {
-      Arcane.logger.log(
-        "User account created successfully but confirmation is required.",
-        level: Level.info,
-      );
-    }
-
     return result;
   }
 
-  /// Confirms the user's signup using their [email] and [confirmationCode].
-  /// Returns a [Result.ok(true)] if signup was successful.
+  /// Confirms the user's signup using their `email` and `confirmationCode`.
+  /// Returns a `Result.ok(true)` if signup was successful.
   Future<Result<bool, String>> confirmSignup({
     required String email,
     required String confirmationCode,
   }) async {
-    if (status == AuthenticationStatus.debug) {
-      return Result.error("Operation not supported in debug mode.");
-    }
-
+    if (_mocked) return Result.ok(false);
     final Result<bool, String> result = await authInterface.confirmSignup(
       username: email,
       confirmationCode: confirmationCode,
@@ -200,28 +189,22 @@ class ArcaneAuthenticationService extends ArcaneService {
   /// Re-sends a verification code to be used when confirming the user's
   /// registration.
   Future<Result<String, String>> resendVerificationCode(String email) async {
-    if (ArcaneAuthenticationService.I.status == AuthenticationStatus.debug) {
-      return Result.error("Operation not supported in debug mode.");
-    }
-
+    if (_mocked) return Result.ok("");
     return authInterface.resendVerificationCode(email);
   }
 
-  /// Attempts to reset the user's password using their [email]. This method
+  /// Attempts to reset the user's password using their `email`. This method
   /// should be called twice. The first call will initialize the password reset
-  /// process. In the first attempt, only the [email] is provided. The second
-  /// call should include the [email], as well as a [newPassword] and
-  /// [confirmationCode]. If the second call is successful, the password will be
+  /// process. In the first attempt, only the `email` is provided. The second
+  /// call should include the `email`, as well as a `newPassword` and
+  /// `confirmationCode`. If the second call is successful, the password will be
   /// reset.
   Future<Result<bool, String>> resetPassword({
     required String email,
     String? newPassword,
     String? confirmationCode,
   }) async {
-    if (status == AuthenticationStatus.debug) {
-      return Result.error("Operation not supported in debug mode.");
-    }
-
+    if (_mocked) return Result.ok(false);
     final Result<bool, String> result = await authInterface.resetPassword(
       email: email,
       newPassword: newPassword,
@@ -231,98 +214,11 @@ class ArcaneAuthenticationService extends ArcaneService {
     return result;
   }
 
+  /// Marks the service as mocked for testing purposes.
+  ///
+  /// If the service is mocked, no method will be executed.
   @visibleForTesting
   static void setMocked() {
     _mocked = true;
-  }
-}
-
-enum AuthenticationStatus {
-  authenticated,
-  unauthenticated,
-  debug,
-  ;
-
-  bool get isDebug => this == debug;
-  bool get isAuthenticated => this == authenticated;
-  bool get isUnauthenticated => this == unauthenticated;
-}
-
-abstract class ArcaneAuthInterface {
-  /// Returns true if the user is signed in.
-  Future<bool> get isSignedIn;
-
-  /// Returns the access token.
-  Future<String?> get accessToken;
-
-  /// Returns the refresh token.
-  Future<String?> get refreshToken;
-
-  /// Initializes the auth interface
-  Future<void> init();
-
-  /// Logs the user out.
-  Future<Result<void, String>> logout();
-
-  /// Logs the user in using an email address and password.
-  Future<Result<void, String>> loginWithEmailAndPassword({
-    required String email,
-    required String password,
-  });
-
-  /// Re-sends a verification code to the user's [email] address.
-  Future<Result<String, String>> resendVerificationCode(
-    String email,
-  );
-
-  /// Signs a user up with a username, password, and email.
-  Future<Result<SignUpStep, String>> signup({
-    required String password,
-    required String email,
-  });
-
-  /// Confirms a user's signup using an [username] and the [confirmationCode]
-  /// they received from the signup process.
-  Future<Result<bool, String>> confirmSignup({
-    required String username,
-    required String confirmationCode,
-  });
-
-  /// Resets a user's password using an [email] address and the [code] they
-  /// received from the reset password process.
-  Future<Result<bool, String>> resetPassword({
-    required String email,
-    String? newPassword,
-    String? code,
-  });
-}
-
-enum SignUpStep {
-  confirmSignUp,
-  done,
-}
-
-enum Environment { debug, normal }
-
-class ArcaneEnvironment extends Cubit<Environment> {
-  ArcaneEnvironment() : super(Environment.normal);
-
-  Future<void> enableDebugMode(Future<void> Function()? onDemoModeSet) async {
-    if (onDemoModeSet != null) await onDemoModeSet();
-
-    emit(Environment.debug);
-  }
-}
-
-class ArcaneEnvironmentProvider extends StatelessWidget {
-  final Widget child;
-  const ArcaneEnvironmentProvider({required this.child, super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocProvider(
-      create: (context) => ArcaneEnvironment(),
-      child: child,
-    );
   }
 }
